Added dry run to image prune.
* Update: Still work-in-progress
Image prunings in dry run mode were not reported when `docker system
prune` was called because the space reclaimed was always reported to be
0B. In order to address this, I've made some modifications and left in
debug statements. However, I still have not been able to get this to
work.

* Oldest commit message:
This is a work-in-progress commit.
While image pruning in dry run reports the images without deleting them,
the estimated space reclaimed is reported incorrectly (as 0B always).
This implies that `docker image prune --dry-run --all` shows what images
will be deleted but the space reclaimed is reported as 0B.
Furthermore, `docker system prune --dry-run --all` will not even report
what images will be deleted because in system/prune.go output is
suppressed for prune reports that have space reclaimed as 0B.
(See line 91 in system/prune.go)